Added Vacuum project files.
Project files for the vacuum project. Basically determines the state of 4 squares, whether or not they are dirty or clean and attempts to clean dirty spots. The vacuum can only occupy one square at a time and has a chance to fail its action on a square and might clean a square that is not dirty or fail to clean a dirty square. Squares become dirty randomly during each step. A step consists of a turn move where the vacuum can either move or attempt to clean a spot. The vacuum will attempt to determine which squares are dirty at the start of the turn but has a chance to be wrong for every square.

Outputs and output layout are incorrect. Will have to fix at a later date.