Added Maze project files
This program solves mazes by taking in square mazes via text files and outputting a series if directions to take from the starting point to reach the goal in the shortest step cost. Each step cost is denotated as one space from the current space. Spaces at the border of the maze cost an additional 10 cost, so 11 total. Utilizes BFs, DFS, and UCS. 

Working as intended. Code complete!