Added the files we edited for our AI Game project
Basically, we (3 students) edited these files from the project file in order to get MiniMax working then we made the AlphaBetaBreakthough file that implemented alphabeta searching based on heuristics we made. Based on our heuristics, it would make moves on several temporary boards found through recursion and evaluate the score of that board based on our heuristic. For example, if we had a board where our piece was about to win and it wasn't threatened, then it would choose to make that move unless a better one was found. I was in charge of getting the Minimax and AlphabetaSearch up and running while my 2 partners worked on the heuristics and other features such as an opening book of moves. 

The actual project was too big for github. I'll see if I can upload the actual project in some other way.